+++ mp-opt-model/CHANGES.md
@@ -5,6 +5,12 @@ Change history for MP-Opt-Model
 Changes since 4.0
 -----------------
 
+#### 10/27/22
+  - Add `runtime` field to `output` argument of `qps_glpk()` and
+    `qps_mosek()`.
+  - Add support to `qps_master()` for calling `qps_<my_solver>()` by setting
+    `opt.alg` to `'<MY_SOLVER>'` to allow for custom solvers.
+
 #### 7/5/22
   - Update for compatibility with Artelys Knitro 13.1 and later.
 
@@ -75,7 +81,7 @@ Version 4.0 - *Oct 18, 2021*
 #### 12/16/20
   - Update to use labels from `set_types` property as headers for
     `opt_model/display()` to simplify things and facilitate use by
-    sub-classes.
+    subclasses.
 
 #### 12/1/20
   - Move `init_set_types()` call out of `opt_model` constructor to avoid
@@ -281,7 +287,7 @@ Version 0.8 - *Apr 29, 2020*
     object appears first. Ideally, these would be defined as methods
     of the `opt_model` class, but Octave 4.2 and earlier is not
     able to find them via a function handle (as used in the `solve()`
-    method) if they are inherited by a sub-class.
+    method) if they are inherited by a subclass.
     - `opf_consfcn()` --> `nlp_consfcn()`
     - `opf_costfcn()` --> `nlp_costfcn()`
     - `opf_hessfcn()` --> `nlp_hessfcn()`